feat(protocol-designer): move redux types from shared types into sub modules
forgot to do this in the last redux infra PR!
